NEXT-20525 - Implement app context change

See merge request shopware/6/product/platform!7743

+++ src/Administration/Resources/app/administration/src/app/init/extension-data-handling.init.ts
@@ -1,4 +1,32 @@
 import type { Entity } from '@shopware-ag/admin-extension-sdk/es/data/_internals/Entity';
+import type EntityCollection from '../../core/data/entity-collection.data';
+import type Repository from '../../core/data/repository.data';
+
+function getRepository(entityName: string, additionalInformation: { _event_: MessageEvent<string>}): Repository | null {
+    const extensionName = Object.keys(Shopware.State.get('extensions'))
+        .find(key => Shopware.State.get('extensions')[key].baseUrl.startsWith(additionalInformation._event_.origin));
+
+    if (!extensionName) {
+        return null;
+    }
+
+    const extension = Shopware.State.get('extensions')?.[extensionName];
+    if (!extension) {
+        return null;
+    }
+
+    if (extension.integrationId) {
+        return Shopware.Service('repositoryFactory')
+            .create(entityName, '', { 'sw-app-integration-id': extension.integrationId });
+    }
+
+    return Shopware.Service('repositoryFactory').create(entityName);
+}
+
+function rejectRepositoryCreation(entityName: string): unknown {
+    // eslint-disable-next-line prefer-promise-reject-errors
+    return Promise.reject(`Could not create repository for entity "${entityName}"`);
+}
 
 export default function initializeExtensionDataLoader(): void {
     Shopware.ExtensionAPI.handle('repositorySearch', (
@@ -7,11 +35,15 @@ export default function initializeExtensionDataLoader(): void {
             criteria = new Shopware.Data.Criteria(),
             context,
         },
+        additionalInformation,
     ) => {
-        const repository = Shopware.Service('repositoryFactory').create(entityName);
+        const repository = getRepository(entityName, additionalInformation);
+        if (!repository) {
+            return rejectRepositoryCreation(entityName) as Promise<EntityCollection>;
+        }
+
         const mergedContext = { ...Shopware.Context.api, ...context };
 
-        // TODO NEXT-20525: add header parameter for app privileges context
         return repository.search(criteria, mergedContext);
     });
 
@@ -22,11 +54,15 @@ export default function initializeExtensionDataLoader(): void {
             criteria = new Shopware.Data.Criteria(),
             context,
         },
+        additionalInformation,
     ) => {
-        const repository = Shopware.Service('repositoryFactory').create(entityName);
+        const repository = getRepository(entityName, additionalInformation);
+        if (!repository) {
+            return rejectRepositoryCreation(entityName) as Promise<null>;
+        }
+
         const mergedContext = { ...Shopware.Context.api, ...context };
 
-        // TODO NEXT-20525: add header parameter for app privileges context
         return repository.get(id, mergedContext, criteria);
     });
 
@@ -36,11 +72,15 @@ export default function initializeExtensionDataLoader(): void {
             entity,
             context,
         },
+        additionalInformation,
     ) => {
-        const repository = Shopware.Service('repositoryFactory').create(entityName);
+        const repository = getRepository(entityName, additionalInformation);
+        if (!repository) {
+            return rejectRepositoryCreation(entityName) as Promise<void>;
+        }
+
         const mergedContext = { ...Shopware.Context.api, ...context };
 
-        // TODO NEXT-20525: add header parameter for app privileges context
         return repository.save(entity, mergedContext) as Promise<void>;
     });
 
@@ -51,11 +91,15 @@ export default function initializeExtensionDataLoader(): void {
             entityId,
             context,
         },
+        additionalInformation,
     ) => {
-        const repository = Shopware.Service('repositoryFactory').create(entityName);
+        const repository = getRepository(entityName, additionalInformation);
+        if (!repository) {
+            return rejectRepositoryCreation(entityName);
+        }
+
         const mergedContext = { ...Shopware.Context.api, ...context };
 
-        // TODO NEXT-20525: add header parameter for app privileges context
         return repository.clone(entityId, mergedContext, behavior as $TSDangerUnknownObject);
     });
 
@@ -64,10 +108,13 @@ export default function initializeExtensionDataLoader(): void {
             entityName,
             entity,
         },
+        additionalInformation,
     ) => {
-        const repository = Shopware.Service('repositoryFactory').create(entityName);
+        const repository = getRepository(entityName, additionalInformation);
+        if (!repository) {
+            return rejectRepositoryCreation(entityName) as Promise<boolean>;
+        }
 
-        // TODO NEXT-20525: add header parameter for app privileges context
         return repository.hasChanges(entity);
     });
 
@@ -77,11 +124,15 @@ export default function initializeExtensionDataLoader(): void {
             entities,
             context,
         },
+        additionalInformation,
     ) => {
-        const repository = Shopware.Service('repositoryFactory').create(entityName);
+        const repository = getRepository(entityName, additionalInformation);
+        if (!repository) {
+            return rejectRepositoryCreation(entityName)as Promise<void>;
+        }
+
         const mergedContext = { ...Shopware.Context.api, ...context };
 
-        // TODO NEXT-20525: add header parameter for app privileges context
         return repository.saveAll(entities, mergedContext) as Promise<void>;
     });
 
@@ -91,11 +142,15 @@ export default function initializeExtensionDataLoader(): void {
             entityId,
             context,
         },
+        additionalInformation,
     ) => {
-        const repository = Shopware.Service('repositoryFactory').create(entityName);
+        const repository = getRepository(entityName, additionalInformation);
+        if (!repository) {
+            return rejectRepositoryCreation(entityName)as Promise<void>;
+        }
+
         const mergedContext = { ...Shopware.Context.api, ...context };
 
-        // TODO NEXT-20525: add header parameter for app privileges context
         return repository.delete(entityId, mergedContext) as unknown as Promise<void>;
     });
 
@@ -105,11 +160,15 @@ export default function initializeExtensionDataLoader(): void {
             entityId,
             context,
         },
+        additionalInformation,
     ) => {
-        const repository = Shopware.Service('repositoryFactory').create(entityName);
+        const repository = getRepository(entityName, additionalInformation);
+        if (!repository) {
+            return rejectRepositoryCreation(entityName) as Promise<Entity>;
+        }
+
         const mergedContext = { ...Shopware.Context.api, ...context };
 
-        // TODO NEXT-20525: add header parameter for app privileges context
         return repository.create(mergedContext, entityId) as Entity;
     });
 }
